@@ -1,11 +1,3 @@
-apply plugin: 'net.minecraftforge.gradle.forge'
-
-apply from: 'gradle/scripts/common.gradle'
-apply from: 'gradle/scripts/with-at.gradle'
-apply from: 'gradle/scripts/with-api.gradle'
-
-apply from: 'dependencies.gradle'
-
 buildscript {
     repositories {
         jcenter()
@@ -15,14 +7,23 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'net.minecraftforge.gradle:ForgeGradle:2.2-SNAPSHOT'
+        classpath 'net.minecraftforge.gradle:ForgeGradle:2.3-SNAPSHOT'
     }
 }
+apply plugin: 'net.minecraftforge.gradle.forge'
+
+apply from: 'gradle/scripts/common.gradle'
+apply from: 'gradle/scripts/with-at.gradle'
+apply from: 'gradle/scripts/with-api.gradle'
+
+apply from: 'dependencies.gradle'
 
 def BUILD_NUMBER = System.getenv("BUILD_NUMBER") ?: "0"
 
-sourceCompatibility = JavaVersion.VERSION_1_7
-targetCompatibility = JavaVersion.VERSION_1_7
+sourceCompatibility = targetCompatibility = '1.8'
+compileJava {
+    sourceCompatibility = targetCompatibility = '1.8'
+}
 
 version = major_version + "." + minor_version + "." + BUILD_NUMBER
 group = "net.blay09.mods"
@@ -31,7 +32,7 @@ archivesBaseName = mod_name + "_" + minecraft_version
 minecraft {
     version = minecraft_version + "-" + forge_version
     runDir = "run"
-    useDepAts = false
+    //useDepAts = false
     mappings = mappings_version
 }
 
