@@ -17,12 +17,9 @@ public class HandlerCreateCowJar implements IMessageHandler<MessageCreateCowJar,
 	@Override
 	@Nullable
 	public IMessage onMessage(final MessageCreateCowJar message, MessageContext ctx) {
-		NetworkHandler.getThreadListener(ctx).addScheduledTask(new Runnable() {
-			@Override
-			public void run() {
-				Minecraft.getMinecraft().world.playSound(message.getPos().getX(), message.getPos().getY(), message.getPos().getZ(), SoundEvents.ENTITY_CHICKEN_EGG, SoundCategory.BLOCKS, 1f, 1f, false);
-				Minecraft.getMinecraft().world.spawnParticle(EnumParticleTypes.EXPLOSION_LARGE, true, message.getPos().getX(), message.getPos().getY() + 1, message.getPos().getZ(), 0f, 0f, 0f);
-			}
+		NetworkHandler.getThreadListener(ctx).addScheduledTask(() -> {
+			Minecraft.getMinecraft().world.playSound(message.getPos().getX(), message.getPos().getY(), message.getPos().getZ(), SoundEvents.ENTITY_CHICKEN_EGG, SoundCategory.BLOCKS, 1f, 1f, false);
+			Minecraft.getMinecraft().world.spawnParticle(EnumParticleTypes.EXPLOSION_LARGE, true, message.getPos().getX(), message.getPos().getY() + 1, message.getPos().getZ(), 0f, 0f, 0f);
 		});
 		return null;
 	}
