@@ -16,14 +16,11 @@ public class HandlerItemList implements IMessageHandler<MessageItemList, IMessag
     @Override
 	@Nullable
     public IMessage onMessage(final MessageItemList message, MessageContext ctx) {
-		NetworkHandler.getThreadListener(ctx).addScheduledTask(new Runnable() {
-			@Override
-			public void run() {
-				Container container = FMLClientHandler.instance().getClientPlayerEntity().openContainer;
-				if(container instanceof ContainerRecipeBook) {
-					((ContainerRecipeBook) container).setItemList(message.getRecipeList());
-					((ContainerRecipeBook) container).setHasOven(message.getHasOven());
-				}
+		NetworkHandler.getThreadListener(ctx).addScheduledTask(() -> {
+			Container container = FMLClientHandler.instance().getClientPlayerEntity().openContainer;
+			if(container instanceof ContainerRecipeBook) {
+				((ContainerRecipeBook) container).setItemList(message.getRecipeList());
+				((ContainerRecipeBook) container).setHasOven(message.getHasOven());
 			}
 		});
         return null;
