@@ -1,8 +1,7 @@
-package net.blay09.mods.cookingforblockheads.blaycommon;
+package net.blay09.mods.cookingforblockheads;
 
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.items.IItemHandler;
