@@ -7,6 +7,7 @@ import net.minecraft.block.material.Material;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.resources.I18n;
+import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Items;
 import net.minecraft.item.EnumDyeColor;
@@ -14,15 +15,20 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
+import javax.annotation.Nullable;
 import java.util.List;
 
 public class BlockCorner extends BlockKitchen {
+
+	public static final String name = "corner";
+	public static final ResourceLocation registryName = new ResourceLocation(CookingForBlockheads.MOD_ID, name);
 
 	private static final AxisAlignedBB[] BOUNDING_BOXES = new AxisAlignedBB[] {
 			new AxisAlignedBB(0, 0, 0, 0.96875, 0.9375, 0.96875),
@@ -34,8 +40,7 @@ public class BlockCorner extends BlockKitchen {
 	public BlockCorner() {
 		super(Material.ROCK);
 
-		setRegistryName(CookingForBlockheads.MOD_ID, "corner");
-		setUnlocalizedName(getRegistryNameString());
+		setUnlocalizedName(registryName.toString());
 		setSoundType(SoundType.STONE);
 		setHardness(5f);
 		setResistance(10f);
@@ -93,9 +98,9 @@ public class BlockCorner extends BlockKitchen {
 	}
 
 	@Override
-	public void addInformation(ItemStack stack, EntityPlayer player, List<String> tooltip, boolean advanced) {
-		super.addInformation(stack, player, tooltip, advanced);
-		for (String s : I18n.format("tooltip." + getRegistryName() + ".description").split("\\\\n")) {
+	public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag advanced) {
+		super.addInformation(stack, world, tooltip, advanced);
+		for (String s : I18n.format("tooltip." + registryName + ".description").split("\\\\n")) {
 			tooltip.add(TextFormatting.GRAY + s);
 		}
 	}
