@@ -1,13 +1,18 @@
 package net.blay09.mods.cookingforblockheads.block;
 
+import net.blay09.mods.cookingforblockheads.CookingForBlockheads;
 import net.blay09.mods.cookingforblockheads.tile.TileCowJar;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
 
 public class BlockCowJar extends BlockMilkJar {
 
+    public static final String name = "cow_jar";
+    public static final ResourceLocation registryName = new ResourceLocation(CookingForBlockheads.MOD_ID, name);
+
     public BlockCowJar() {
-        super("cow_jar");
+        setUnlocalizedName(registryName.toString());
     }
 
     @Override
