@@ -6,45 +6,40 @@ import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.client.resources.I18n;
-import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.TextFormatting;
+import net.minecraft.world.World;
 import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
+import javax.annotation.Nullable;
 import java.util.List;
 
 public class BlockKitchenFloor extends Block {
 
+	public static final String name = "kitchen_floor";
+	public static final ResourceLocation registryName = new ResourceLocation(CookingForBlockheads.MOD_ID, name);
+
 	public BlockKitchenFloor() {
 		super(Material.ROCK);
 
-		setRegistryName(CookingForBlockheads.MOD_ID, "kitchen_floor");
-		setUnlocalizedName(getRegistryNameString());
+		setUnlocalizedName(registryName.toString());
 		setSoundType(SoundType.STONE);
+		setCreativeTab(CookingForBlockheads.creativeTab);
 		setHardness(5f);
 		setResistance(10f);
-		setCreativeTab(CookingForBlockheads.creativeTab);
 	}
 
 	@Override
-	public void addInformation(ItemStack stack, EntityPlayer player, List<String> tooltip, boolean advanced) {
+	public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag advanced) {
 		tooltip.add(TextFormatting.YELLOW + I18n.format("tooltip." + CookingForBlockheads.MOD_ID + ":multiblock_kitchen"));
-		for (String s : I18n.format("tooltip." + getRegistryName() + ".description").split("\\\\n")) {
+		for (String s : I18n.format("tooltip." + registryName + ".description").split("\\\\n")) {
 			tooltip.add(TextFormatting.GRAY + s);
 		}
-	}
-
-	@SideOnly(Side.CLIENT)
-	public void registerModels() {
-		ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(this), 0, new ModelResourceLocation(getRegistryNameString(), "inventory"));
-	}
-
-	public String getRegistryNameString() {
-		//noinspection ConstantConditions
-		return getRegistryName().toString();
 	}
 
 }
