@@ -1,13 +1,14 @@
 package net.blay09.mods.cookingforblockheads.block;
 
 import net.blay09.mods.cookingforblockheads.CookingForBlockheads;
-import net.blay09.mods.cookingforblockheads.blaycommon.ItemUtils;
+import net.blay09.mods.cookingforblockheads.ItemUtils;
 import net.blay09.mods.cookingforblockheads.network.handler.GuiHandler;
 import net.blay09.mods.cookingforblockheads.tile.TileSpiceRack;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.resources.I18n;
+import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemBlock;
@@ -15,6 +16,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextFormatting;
@@ -26,6 +28,9 @@ import java.util.List;
 
 public class BlockSpiceRack extends BlockKitchen {
 
+	public static final String name = "spice_rack";
+	public static final ResourceLocation registryName = new ResourceLocation(CookingForBlockheads.MOD_ID, name);
+
 	private static final AxisAlignedBB[] BOUNDING_BOXES = new AxisAlignedBB[] {
 			new AxisAlignedBB(0, 0.25, 1 - 0.125, 1, 1, 1),
 			new AxisAlignedBB(0, 0.25, 0, 1, 1, 0.125),
@@ -36,8 +41,7 @@ public class BlockSpiceRack extends BlockKitchen {
     public BlockSpiceRack() {
         super(Material.WOOD);
 
-		setRegistryName(CookingForBlockheads.MOD_ID, "spice_rack");
-		setUnlocalizedName(getRegistryNameString());
+		setUnlocalizedName(registryName.toString());
         setSoundType(SoundType.WOOD);
         setHardness(2.5f);
     }
@@ -136,10 +140,10 @@ public class BlockSpiceRack extends BlockKitchen {
 		super.breakBlock(world, pos, state);
 	}
 
-    @Override
-    public void addInformation(ItemStack stack, EntityPlayer player, List<String> tooltip, boolean advanced) {
-        super.addInformation(stack, player, tooltip, advanced);
-        for (String s : I18n.format("tooltip." + getRegistryName() + ".description").split("\\\\n")) {
+	@Override
+	public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag advanced) {
+		super.addInformation(stack, world, tooltip, advanced);
+        for (String s : I18n.format("tooltip." + registryName + ".description").split("\\\\n")) {
             tooltip.add(TextFormatting.GRAY + s);
         }
     }
