@@ -20,7 +20,7 @@ public class SortButtonName implements ISortButton {
     
     @Override
     public String getTooltip() {
-        return "tooltip." + CookingForBlockheads.MOD_ID + ":sort_by_name";
+        return "tooltip.cookingforblockheads:sort_by_name";
     }
     
     @Override
