@@ -24,7 +24,7 @@ public class FakeSlotRecipe extends FakeSlot {
 	}
 
 	@Override
-	public boolean canBeHovered() {
+	public boolean isEnabled() {
 		return recipe != null;
 	}
 
