@@ -3,8 +3,10 @@ package net.blay09.mods.cookingforblockheads.tile;
 import net.blay09.mods.cookingforblockheads.api.capability.CapabilityKitchenItemProvider;
 import net.blay09.mods.cookingforblockheads.api.capability.KitchenItemProvider;
 import net.blay09.mods.cookingforblockheads.block.BlockCounter;
+import net.blay09.mods.cookingforblockheads.compat.Compat;
 import net.blay09.mods.cookingforblockheads.network.VanillaPacketHandler;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.SoundEvents;
 import net.minecraft.item.EnumDyeColor;
 import net.minecraft.nbt.NBTTagCompound;
@@ -20,11 +22,11 @@ import net.minecraftforge.fml.common.Optional;
 import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.ItemStackHandler;
-import vazkii.quark.base.handler.IDropoffManager;
+import vazkii.quark.api.IDropoffManager;
 
 import javax.annotation.Nullable;
 
-@Optional.Interface(modid = "Quark", iface = "vazkii.quark.base.handler.IDropoffManager", striprefs = true)
+@Optional.Interface(modid = Compat.QUARK, iface = "vazkii.quark.api.IDropoffManager", striprefs = true)
 public class TileCounter extends TileEntity implements ITickable, IDropoffManager {
 
     private final ItemStackHandler itemHandler = new ItemStackHandler(27) {
@@ -172,8 +174,8 @@ public class TileCounter extends TileEntity implements ITickable, IDropoffManage
     }
 
     @Override
-    @Optional.Method(modid = "Quark")
-    public boolean acceptsDropoff() {
+    public boolean acceptsDropoff(EntityPlayer entityPlayer) {
         return true;
     }
+
 }
